vector_db/requirements.txt
# Core Dependencies
fastapi>=0.109.0
uvicorn>=0.27.0
python-dotenv>=1.0.0
pydantic>=2.5.0

# Vector Database
faiss-cpu>=1.7.4
numpy>=1.24.0
sentence-transformers>=2.2.2

# Storage & Caching
redis>=5.0.1

# Monitoring & Logging
prometheus-client>=0.19.0
python-json-logger>=2.0.7

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
httpx>=0.26.0 